# [PERF] hr_recruitment: add index on `active`