@robodoo 2024-06-27
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/171000.png)](https://mergebot.odoo.com/odoo/odoo/pull/171000)