@duong77476-viindoo 2024-07-02
I wonder we can do this for other large table like task