@pivi-odoo 2024-07-02
Hello @duong77476-viindoo 👋🏼,

It would depends on the workflow for the model in question. Here applicants are usually *archived* when we don't need them anymore. AFAIK tasks on the other hand usually have just their state changed to Done/Canceled once they are moved to an ending stage, but are rarely archived themselves. So for tasks, adding an index on `active` would be useless.