@duong77476-viindoo 2024-07-02
Thanks @pivi-odoo